updated xmla_endpoint
I updated the xmla_endpoint in the "arg" section so that it would get the value from the "workspace" variable

This came about when I was testing the notebook and it was failing because the semantic model was in another workspace. So the xmla_endpoint was not being passed through to the correct endpoint.